Clean the code in map
MapUtils.kt: Put some utils function in it
Map.kt : Move some function in MapUtils.kt
ConstantMap.kt: Add some constant
MapClusteringTest.kt : Add constant
MapClusterImageTest.kt.kt : Add constant
